feat: prepare types
