--- src/types/product.ts
@@ -1,8 +1,0 @@
-export type Product = {
-  imagePath: string
-  title: string
-  desc: string
-  url: string
-  github: string
-  techs: string[]
-}
